# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "InfyOm Laravel Generator APIs",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/house-residents": {
            "get": {
                "tags": [
                    "HouseResident"
                ],
                "summary": "getHouseResidentList",
                "description": "Get all HouseResidents",
                "operationId": "9d546d7126127f67b32c46ae185d7f1f",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/HouseResident"
                                            }
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "HouseResident"
                ],
                "summary": "createHouseResident",
                "description": "Create HouseResident",
                "operationId": "e732b5e367ea64ea2f39452fa8976699",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/HouseResident"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/HouseResident"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/house-residents/{id}": {
            "get": {
                "tags": [
                    "HouseResident"
                ],
                "summary": "getHouseResidentItem",
                "description": "Get HouseResident",
                "operationId": "f24b528fec335252f26c545be235f3d4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of HouseResident",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/HouseResident"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "HouseResident"
                ],
                "summary": "updateHouseResident",
                "description": "Update HouseResident",
                "operationId": "dd3a81444e3be9a010ae76c143fd04e7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of HouseResident",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/HouseResident"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/HouseResident"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "HouseResident"
                ],
                "summary": "deleteHouseResident",
                "description": "Delete HouseResident",
                "operationId": "9f5a2607139cc060b9255345b21cd3b7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of HouseResident",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "string"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "HouseResident": {
                "required": [
                    "house_id",
                    "resident_id",
                    "start_date"
                ],
                "properties": {
                    "start_date": {
                        "description": "",
                        "type": "string",
                        "format": "date",
                        "readOnly": false,
                        "nullable": false
                    },
                    "end_date": {
                        "description": "",
                        "type": "string",
                        "format": "date",
                        "readOnly": false,
                        "nullable": true
                    },
                    "created_at": {
                        "description": "",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true,
                        "nullable": true
                    },
                    "updated_at": {
                        "description": "",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true,
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}